requirements.txt
torch
transformers
peft
accelerate
bitsandbytes
datasets
sentencepiece
streamlit
GPUtil
python-dotenv
wandb